+++ spec/features/beis_users_can_upload_level_b_activities_spec.rb
@@ -317,6 +317,28 @@ RSpec.feature "BEIS users can upload Level B activities" do
         "Comments"
       ])
     end
+
+    scenario "uploading a valid set of activities" do
+      old_count = Activity.count
+
+      within ".upload-form--ispf-non-oda" do
+        attach_file_and_click_submit(filepath: "spec/fixtures/csv/valid_level_b_ispf_non_oda_activities_upload.csv")
+      end
+
+      expect(Activity.count - old_count).to eq(1)
+
+      new_activity = Activity.find_by(title: "A title")
+
+      visit organisation_activities_path(organisation)
+
+      within "//tbody" do
+        expect(page).to have_content(new_activity.title)
+      end
+
+      visit organisation_activity_comments_path(organisation, new_activity)
+
+      expect(page).to have_text("This is a comment")
+    end
   end
 
   def expect_change_to_be_recorded_as_historical_event(